emerge-webrsync: drop the get_unixtime() and get_date_part() functions
Drop both the get_unixtime() and get_date_part() functions in favour of
utilising the %(fmt)T format that is supported by the printf builtin.

BEFORE

$ date -u +%s                       # print time of day in unixtime
$ date -u -d "@${unixtime}" +%Y%m%d # convert unixtime to a %Y%m%d date

AFTER

$ printf '%(%s)T\n'
$ printf '%(%Y%m%d)T\n' "${unixtime}"

Note that, as of bash 4.3, no arguments need be specified after the
format string in order to obtain the current time of day. Older versions
required an argument of -1 for that purpose.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>