emerge-webrsync: don't check whether gpg(1) exists in check_file_signature_gpg_unwrapped()
As concerns the check_file_signature_gpg_unwrapped() function, refrain
from checking whether the gpg(1) utility can be found in PATH. This is
what I would refer to as a pointless proxy check. One need not care
whether gpg(1) can be found in PATH. For if it cannot be found then the
first invocation of it shall fail, in which case, so be it. Let it fail!

Having dispensed with the proxy check, it can be seen that the
diagnostic message raised by the shell is sufficiently clear.

 * Checking digest ...
 * Checking signature ...
 * Falling back to gpg as gemato is not installed
/usr/bin/emerge-webrsync: line 335: gpg: command not found

I would add that proxy checks are loved by programmers who have a
tendency to avoid thinking about the matter of error handling in
general, exhibiting no diligence whatsover in this respect. Whensoever I
encounter such checks, I am immediately inclined to regard them as a
potential red flag. As a case in point, prior to my working on the
emerge-webrsync utility, the very first invocation of gpg(1) in this
function was bereft of an error check. For that matter, just because
type -P indicates that a given utility exists in PATH, it is by no means
guaranteed that it exists at the point of attempted execution.

Now, testing for the existence of a required utility can be a reasonable
thing to do before undergoing any substantive work and thus wasting the
user's time. Yet, had that genuinely been the intent, the check in
question would not have been situated within this particular function to
begin with.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>