emerge-webrsync: don't check for directory existence independently of mkdir -p
If only mkdir(1) had an option that would instruct it not to propagate
EEXIST errors. That way, one would be able to avoid an independent test
for the existence of the directory that amounts to a minor TOCTOU. One
might name the option -p, perhaps. What say you, dear reader?

Link: https://en.wikipedia.org/wiki/Time-of-check_to_time-of-use
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>