emerge-webrsync: clarify behaviour where gemato is missing
Presently, the check_file_signature_gemato() function checks whether
gemato is in PATH. In order to find out what happens if the check fails,
it is necessary to read a comparatively long branch of code. Remedy this
by inverting the sense of the test and making it clear that that
behaviour is to return 127 in that case.

See-also: 7dcae1c07118505a324f1c5d4eb1f37ca9df8999
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>