emerge-webrsync: let check_file_signature() be responsible for dying
The check_file_signature() function acts as an adapter that invokes
check_file_signature_gemato() and/or check_file_signature_unwrapped() in
turn. As such, it contains code to check their return values and invoke
die() if necessary. However, this code is never reached because both of
the wrapped functions invoke die() of their own volition.

Address this issue by removing the redundant error-handling code from
both of the wrapped functions and allowing for check_file_signature() to
perform the error handling, as was originally intended.

Additionally, improve the diagnostic message by incorporating the
basename of the file for which validation has failed. The following
sample output depicts this change.

gpg: WARNING: This key is not certified with a trusted signature!
gpg:          There is no indication that the signature belongs to the owner.
Primary key fingerprint: DCD0 5B71 EAB9 4199 527F  44AC DB6B 8C1F 96D8 BF6D
     Subkey fingerprint: E1D6 ABB6 3BFC FB4B A02F  DF1C EC59 0EEA C918 9250
emerge-webrsync: signature verification failed for 'gentoo-20250617.tar.xz'

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>