emerge-webrsync: improve check_file_signature_gpg() diagnostics for missing keys
Presently, the check_file_signature_gpg() function shows the same
diagnostic messages in the event of a missing key, irrespective of
whether PORTAGE_GPG_KEY was defined or not. Disambiguate by showing
distinct messages for both cases.

If PORTAGE_GPG_KEY is unset/empty:

 * Checking signature ...
 * sec-keys/openpgp-keys-gentoo-release does not appear to be installed
emerge-webrsync: '/usr/share/openpgp-keys/gentoo-release.asc' does not exist (or is not a file)

If PORTAGE_GPG_KEY is non-empty:

 * Checking signature ...
 * PORTAGE_GPG_KEY does not appear to have been set correctly
emerge-webrsync: '/var/empty/nonexistent.key' does not exist (or is not a file)

Further, employ the ${param@Q} form of expansion so as to ensure that
all pathnames are legible, and that control characters and (some)
whitespace characters are not printed in their raw form. Doing so is
acceptable, given a target of >=bash-4.4.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>