emerge-webrsync: stop gpg-agent(1) lingering for ephemeral GNUPGHOME dirs
As concerns the check_file_signature_gpg() function, it is possible for
the value of the 'GNUPGHOME' environment variable to be the path of a
temporary directory created by mktemp(1). In that case, the directory is
guaranteed to be removed upon emerge-webrsync exiting. However, a
defunct gpg-agent(1) process will persist for the deleted directory.

Address this issue by declaring the function as a compound command of
the ( … ) form, which declares its own EXIT trap. The trap shall execute
gpgconf(1) in a way that terminates the gpg-agent(1) process that would
otherwise linger.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>