emerge-webrsync: render wget(1) output terse in nature
By default, the wget(1) utility prints a preamble that is seven lines in
length - including a blank line - before rendering its progress bar. It
really is rather distracting.

Render its output terse by passing the --no-verbose option. This option
impedes the interactive progress bar so also pass the --show-progress
option, unless emerge-webrsync was given the --quiet option.

It should be noted that emerge-webrsync already prints an informational
message containing the URI that it is about to fetch, before executing
the chosen user-agent.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>