emerge-webrsync: check gemato availability in handle_pgp_setup()
Presently, the check_file_signature_gemato() returns 127 if it is unable
to find gemato in PATH. In turn, the check_file_signature() function
responds by falling back to legacy gpg verification and calling
check_file_signature_gpg() instead. Though this approach works, it could
stand to be simplified. Do so in the ways described herewith.

As concerns the check_file_signature_gemato() function, refrain from
checking for the existence of gemato. Instead, always try to execute it.

As concerns the check_file_signature() function, call either the
check_file_signature_gemato() or the check_file_signature_php()
functions, but never both in succession.

As concerns the handle_pgp_setup() function, delegate to it the
responsibility of checking for the existence of gemato. Should it be
found not to exist, select the legacy pgp verification method and print
an appropriate warning.

These changes ensure that, where the 'WEBRSYNC_VERIFY_SIGNATURE'
variable is assigned a value of 2, the legacy gpg method shall always be
employed. I would aver that this behaviour is easier to reason with.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>