emerge-webrsync: don't always abort because PORTAGE_GPG_DIR is empty/unset
A recent commit improved upon the way in which the fallback path from
gemato to gpg(1) is implemented. However, it also made it impossible to
use emerge-webrsync in the case that all these conditions hold true:

- PGP verification hasn't been disabled
- the 'FEATURES' variable doesn't contain "webrsync-gpg"
- gemato cannot be found in 'PATH'
- the 'PORTAGE_GPG_DIR' variable is unset or empty

The symptom of this regression is depicted below.

 * app-portage/gemato does not appear to be installed. Falling back to gpg.
 * PGP verification method: legacy gpg
emerge-webrsync: Please set PORTAGE_GPG_DIR in make.conf!

Rectify this defect by re-positioning the test for whether the
'PORTAGE_GPG_DIR' variable is empty, thereby ensuring that it is only
conducted if the 'FEATURES' variable is found to contain "webrsync-gpg".

Also, revise the diagnostic message so as to be more formal and
matter-of-fact in tone.

Fixes: b54fced4a5a8310bc65972de5b284b249c61315b
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>