emerge-webrsync: make die/exit behave properly in check_file_signature_gpg()
A recent commit introduced the ability for emerge-webrsync to ensure
that defunct gpg-agent(1) processes are terminated upon the
check_file_signature_gpg() function returning, provided that it was
tasked with creating an ephemeral keyring. However, it also introduced a
minor regression as concerns error handling and the conveyance of
diagnostic messages. Consider the following as a case in point.

fingerprint=$(gpg_fingerprint '<infrastructure@gentoo.org>') \
|| die "couldn't find a fingerprint for the <infrastructure@gentoo.org> key"

The intention is straightforward: should it prove impossible to obtain
the fingerprint of the applicable signing key then display a diagnostic
message before immediately exiting the program. Yet, because the
definition of the function is a compound command that incurs a subshell,
only the subshell will be exited. The function will then yield to its
caller, check_file_signature(), which responds as follows.

  die "signature verification failed for ${file@Q}"

Only then shall emerge-webrsync exit, resulting in an additional
diagnostic message being unintentionally shown.

 * Checking signature ...
emerge-webrsync: couldn't find a fingerprint for the <infrastructure@gentoo.org> key
emerge-webrsync: signature verification failed for 'gentoo-20250617.tar.xz'

Address this issue in a manner twofold. Firstly, declare the function as
a compound command of the { … } form so as not to incur a subshell.
Secondly, delegate the responsibility for terminating gpg-agent(1) to
the EXIT trap that is already declared by the main() function.

Additionally, given that the trap is already responsible for wiping the
temporary directory, relocate the trap payload to a dedicated function
by the name of cleanup(). This helps to keep the code reasonably tidy.

Fixes: c9147587da34ecf6cd19bf1ed2d0835d3d8c1777
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>