emerge-webrsync: have both signature checking routines employ common diagnostics
Recently, the diagnostics shown by the check_file_signature_gpg()
function in the event of a missing key were improved. Said commit did
not adjust the diagnostics shown by the check_file_signature_gemato()
function. That is, the former shows the new diagnostics whereas the
latter continues to show the old diagnostics.

Thus, one realises that there is an opportunity to unify them in this
regard. Do so by introducing the assign_key() function and delegating to
it the responsibility of testing for - and reacting to - a missing key.

See-also: bf07fbe382a4a04c65dd76f286bcc061ba11728b
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>