emerge-webrsync: use hash to gauge utility availability
Presently, the type builtin is used to check for the availability of the
following utilities.

- bzcat(1)
- gemato
- md5(1)
- md5sum(1)
- tarsync(1)
- xzcat(1)
- zcat(1)

To do so is questionable. While type -P does coerce bash into performing
a PATH search, the manner in which all of these utilities is invoked
still permits for them to be defined as functions.

Instead, use the hash builtin. This has the advantage of pre-warming -
and subsequently consulting - the cache that the shell employs for
Command Search and Execution.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>