emerge-webrsync: rename WEBRSYNC_VERIFY_SIGNATURE to verification_method
Rename the 'WEBRSYNC_VERIFY_SIGNATURE' variable to
'verification_method'. The choice of a lowercase name reduces the
probability of a namespace conflict and clarifies that the variable is
not expected to be defined by the parent process. It also better conveys
the nature of the variable, in so far as its value is used to
distinguish between distinct verification methods.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>