Add RESTRICT=binpkg support
With the continuing success the Gentoo binhost, it becomes important
mark certain packages as unsuitable for the binhost. Typical examples
include acct-user/*, acct-group/*, or virtual/* packages.

This introduces a new RESTRICT key: binpkg. If set by the ebuild,
portage will not build a binary package for it.

Bugs: https://bugs.gentoo.org/542480#c1
Signed-off-by: Florian Schmaus <flow@gentoo.org>