emerge-webrsync: employ a single method of parsing md5sum digests
Presently, the check_file_digest() function employs two disparate
methods of parsing md5sum digest files.

If the md5sum(1) utility is found to exist, the complete contents of the
file are read and assigned to the 'digest_content' variable, which is
subsequently expanded in a way that strips everything from the first
whitespace character onwards. This method is acceptable, albeit
potentially expensive in the case that the file is larger than expected.

If, on the other hand, the md5(1) utility is found to exist, the cut(1)
utility is used to print the first field of every line in the file. In
doing so, the assumption is made that the file contains only one line.

Address the matter by employing just one method of reading the MD5
checksum from the digest file, which is to use the read builtin to
obtain the first field of the first line. Further, if it cannot then
immediately return false.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>