emerge-werbsync: refactor check_file_digest() and implement diagnostics
This commit aims to improve the check_file_digest() function through
some minor refactoring. The exact changes are described herewith.

Delegate the responsibility of executing the md5sum(1) and md5(1)
utilities to a new function, named md5sum_hex().

In the event that the digest file cannot be opened or read, print a
diagnostic message with ewarn() before returning false, rather than
remain silent.

In the event that the md5sum_hex() function returns false, print a
diagnostic message and exit by calling die(). As before, emerge-webrsync
shall exit if neither of the required utilities can be found. Unlike
before, emerge-webrsync shall also exit if either utility fails, meaning
that it is able to react to severe modes of failure (such as EIO).

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>