emerge-webrsync: reduce the number of arguments required by several functions
Reduce the number of arguments required to just one for all of the
following functions.

- check_file_digest()
- check_file_signature_gemato()
- check_file_signature_gpg
- gpg_verify()

This is achieved by refraining from conveying strings that are easily
composed at the point that they are required. For example, the
check_file_digest() function can compose the name of the digest file by
appending ".md5sum" to the value of the 'file' variable.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>