emerge-webrsync: drop a superfluous informational message from sync_local()
Presently, the sync_local() function displays the "Cleaning up ..."
message after having successfully executed rsync(1). It was once the
case that this message preceded the deletion of the directory containing
the extracted contents of the tarball. However, the applicable rm(1)
command was removed in late 2019.

As such, drop the message. Messages of that kind would now belong in the
cleanup() function and could stand to be more informative.

See-also: 92f4b283c86a74c99fcfa98f8baccaf1b0603dc8
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>