emerge-webrsync: clarify the handling of DISTDIR in main()
This commit aims to clarify the handling of the 'DISTDIR' variable in
the main function(). The exact changes are described herewith.

Refrain from checking the value of the 'keep' key in the 'opt' array
twice.

In the event that the --keep option was not specified, refrain from
executing mkdir(1) with the value of 'DISTDIR' as an operand. Also,
refrain from testing whether the value of 'DISTDIR' is a writable path.
To do either of these things is unnecessary because it is guaranteed
that mktemp(1) will already have succeeded.

In the event that the --keep option was specified, execute mkdir(1) but
also check that it succeeds, exiting the program if it does not.

Employ the ${param@Q} form of expansion in the course of incorporating
the value of 'DISTDIR' into a diagnostic message. Doing so is
acceptable, given a target of >=bash-4.4.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>