emerge-webrsync: purge the directory containing the ephemeral keyring
Presently, the check_file_signature_gpg() function will create an
ephemeral keyring in the case that the 'PORTAGE_GPG_DIR' variable is
unset or empty. However, should that happen, the temporary directory
that is created to contain the keyring is never destroyed. This commit
addresses the issue, while also making some adjacent changes. All of the
changes are described herewith.

As concerns the check_file_signature_gpg() function, refrain from using
the mktemp(1) utility to create the directory that is to contain the
ephemeral keyring. Instead, use mkdir(1) to create a "keyring" directory
that is a sub-directory of the path specified by the 'tmpdir' variable.
Consequently, it is guaranteed to be deleted by the cleanup() function.

As concerns the sync_local() function, if the decision is made to use
tar(1) to extract the tarball then create a "snapshot" directory that is
a sub-directory of the path specified by the 'tmpdir' variable before
switching to it. By extracting the contents of the tarball to this
distinct directory, it remains impossible for rsync(1) to inadvertently
include the contents of the keyring.

As concerns the sync_local() function, refrain from deleting the tarball
in the event that the --keep option was not specified. There is no need
to do so because, where the 'DISTDIR' variable is set to the value of
'tmpdir', it is already guaranteed that the tarball will be removed by
the cleanup() function. Nor are the tarballs particularly large.

As concerns the main() function, execute mktemp(1) in such a way that
the temporary directory is created as an immediate sub-directory of the
path specified by the 'PORTAGE_TMPDIR' variable. That way, it becomes
possible to drop the mkdir -p -- "${PORTAGE_TMPDIR}" command, whose exit
status was not being checked. Also, refrain from switching to the
temporary directory. Owing to the changes made to the sync_local()
function, there is no longer any need to do so.

Fixes: b8ab8e1c850b773dd17e503a22902b52a2d3a868
See-also: c9147587da34ecf6cd19bf1ed2d0835d3d8c1777
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>