emerge-webrsync: drop the obsolete post_sync feature
Presently, the sync_local() function attempts to execute
/etc/portage/bin/post_sync, provided that it is found to be an
executable path. This behaviour has been obsolete since the release of
portage-2.2.16 in Februrary 2015, which introduced support for native
/etc/portage/{repo.,}postsync.d/* hooks.

Given that emerge-webrsync can be executed directly by the user, I
briefly entertained the notion of it being able to execute the
aforementioned hooks without the aid of portage. However, I concluded
that it would be nothing short of dangerous to do so. The fact of the
matter is that emerge-webrsync cannot promise to provide the same
operating environment that portage itself does. As such, it's not
difficult to imagine a user having written a sloppy shell script that
depends on the existence of a particular portage variable in order to
act safely. In the course of discussing the matter with Sam, he agreed
with this assessment.

Thus, drop the code that attempts to execute "post_sync". Users that
wish for postsync hooks to be acted upon should refrain from executing
emerge-webrsync directly and, instead, have the program be executed by
portage by defining "sync-type = webrsync" for the gentoo repo.

Bug: https://bugs.gentoo.org/154220
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>