emerge-webrsync: don't show news if invoked by portage
Presently, the sync_local() function executes emerge(1) for the purpose
of displaying news, provided that the 'FEATURES' variable contains
"news". While it makes sense to do so in the case that the user executes
emerge-webrsync directly, it does not in the case that emerge-webrsync
is executed by portage - as will happen if "sync-type = webrsync" is in
effect for the gentoo repo.

Address the issue by introducing a heuristic by which emerge-webrsync
can determine whether it was invoked by portage, and by refraining from
displaying news in that event.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>