emerge-webrsync: fix exit status reporting for rsync in sync_local()
As concerns the sync_local() function, a recent commit altered the
method by which rsync(1) is detected to have failed, by using the if
command. In many cases, such would be acceptable. However, in this
particular case, the diagnostic message expects to convey the exit
status of the utility, which is clobbered by if.

Address the issue by once again using the "||" control operator in
conjunction with a compound command. Further, improve the quality of the
diagnostic messages, both for eerror() and die().

Fixes: e2dc868b32deace30a470384969cd8f0fe6a1c43
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>