emerge-webrsync: don't delete downloaded files after calling sync_local()
As concerns the do_snapshot() function, refrain from deleting the
downloaded files after having successfully called the sync_local()
function in the event that the --keep option was not specified. There is
no need to do so because, where the 'DISTDIR' variable is set to the
value of 'tmpdir', it is already guaranteed that the files will be
removed by the cleanup() function.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>