isolated-functions.sh: introduce ___parallel() to replace ___parallel_xargs()
Presently, the "isolated-functions.sh" unit provides the
___parallel_xargs() function, whose purpose is to dispatch a simple
command multiple times in parallel, with each invocation taking one
additional parameter. However, in order to be able to issue commands in
parallel, it requires for the xargs(1) utility to support the wholly
non-standard --max-procs option, which only the GNU implementation
provides. This is in spite of the fact that the -P option is commonly
implemented by the current releases of popular unix-like operating
systems, though that, too, is a non-standard option.

This commit introduces a new function, named ___parallel(). It acts in a
way that is very similar to GNU xargs(1), as accompanied by its -0, -L
and -P options. Yet it requires no particular implementation thereof.
Indeed, it does not use xargs(1) at all. Rather, it uses the native
process management features provided by bash. In particular, it takes
advantage of the -n option of the wait builtin, which was introduced by
the release of bash-4.4.

As the following sample commands demonstrate, it is trivial to port any
existing call sites.

# BEFORE
find . -type f -name '*.png' -print0 | ___parallel_xargs -0 pngfix

# AFTER
find . -type f -name '*.png' -print0 | ___parallel pngfix

Indeed, it transpires that ___parallel_xargs() is used only by the
following two units, both of which are ported by this commit.

- bin/ecompress
- bin/install-qa-check.d/60pngfix

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>