phase-functions.sh: support xargs -r across more platforms
Presently, the "isolated-functions.sh" exports a variable by the name of
'XARGS', which is intended to be a string containing both the name of
the ideal xargs(1) implementation and the -r option, if supported.
However, it is implemented in such a way that it may unnecessarily avoid
the use of the -r option. Specifically, it will conclude that the option
is unsupported where both of the following conditions hold true.

- the value of the 'USERLAND' variable is "BSD"
- there is no command available whose name is "gxargs"

It is not a sensible conclusion to draw. Not only is -r a standard
option as of POSIX-1.2024, but various BSD implementations have
supported it for many years. For example, OpenBSD has supported it since
3.8 (2005), FreeBSD since 7.1 (2009) and NetBSD since 5.2 (2012).

Address the issue by first checking whether gxargs is available. If it
is, assume it to be GNU xargs(1). Otherwise, perform a test that
empirically determines whether the option is supported.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>