misc-function.sh: check for tar --xattrs support in a quieter way
Presently, the __dyn_package() function employs the combination of a
conditional expression and a command substitution to test whether tar(1)
supports the --xattrs option. However, should the xtrace shell option
happen to be enabled, the captured output of tar --help will be shown in
its entirety, which is somewhat grating. Address this issue by instead
using grep(1) to match against the output stream of tar(1).

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>