phase-helpers.sh: check for GNU ar in a quieter way
Presently, the unpack() function employs the combination of a
conditional expression and a command substitution to test whether the
implementation of ar(1) is provided by GNU. However, the captured output
of ar --version will be shown in its entirety, provided that the
following two conditions hold true.

- the xtrace shell option is enabled
- BASH_XTRACEFD is set to something other than 2

Specifically, the captured output will be written to the file descriptor
that is referenced by 'BASH_XTRACEFD', which could be somewhat grating.
Address this issue by instead using grep(1) to match against the output
stream of ar(1).

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>