phase-functions.sh: use find(1) to potentially delete PORTAGE_BUILDDIR
Presently, the __dyn_clean() function concludes with an invocation of
rmdir(1) that attempts to delete the directory specified by the
'PORTAGE_BUILDDIR' variable, immediately followed by an invocation of
the true builtin. Clearly, the author of this code acknowledged that the
directory might still be populated, but chose to silence the ensuing
error and jettison the exit status of rmdir(1).

Address this issue by instead invoking find(1) with the primaries
necessary for it to identify an empty directory, and by having it
execute rmdir(1) in that event. Further, pass the "--" operand to
rmdir(1) to signify end-of-options.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>