save-ebuild-env-sh: refrain from assigning to and unsetting x
Presently, the __save_ebuild_env() function uses the 'x' variable to
iterate over various function names before proceeding to unset it. To do
so is inadvisable because variables are dynamically scoped in bash and
there may be no guarantee that 'x' wasn't already declared. Hence, the
use of the unset builtin can be dangerous, since it can have the effect
of unmasking a variable from a previous scope.

$ var=1
$ x() { local var=2; y; }
$ y() { var=3; unset var; declare -p var; }
$ x
declare -- var="1"

Address this issue by instead having the applicable for loop use the
special '_' parameter as a topic variable.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>