phase-functions.sh: use array variables in __filter_readonly_variables()
Presently, the __filter_readonly_variables() function declares the
following scalar variables, each of which is intended to be comprised by
one or more whitespace-separated words.

- bash_misc_vars
- binpkg_untrusted_vars
- filtered_sandbox_vars
- misc_garbage_vars
- readonly_bash_vars

Instad, declare them as array variables and operate upon them
accordingly. This change, along with some minor reformatting, makes for
an improvement in code legibility. It also guarantees that the argument
passed to the filter-bash-environment.py script will be comprised by
words separated only by a single <space>, rather than varying degrees of
horizontal and vertical whitespace, as was the case before.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>