phase-functions.sh: filter bash vars comprehensively in __filter_readonly_variables()
Presently, the __filter_readonly_variables() function attempts to filter
variables that pertain to bash. That is, those which are either set by
bash, or which are otherwise able to influence the behaviour of bash. Up
until recently, a distinction was made between variables known to be
readonly and those that are not. However, it can be demonstrated that
both of the implied sets have been poorly maintained. Consider the prior
definition of the 'readonly_bash_vars', which was as follows.

BASHOPTS BASHPID DIRSTACK EUID FUNCNAME GROUPS PIPESTATUS PPID SHELLOPTS
UID

Ten distinct names. Yet, a new instance of bash in its capacity as a
non-interactive shell defines only six readonly variables.

$ env -i bash -c 'compgen -A variable |
  while read -r; do [[ ${!REPLY@a} == *r* ]] && printf .; done' |
  wc -c
6

Ergo, it was incorrect to categorise 'DIRSTACK', 'FUNCNAME', 'GROUPS'
and 'PIPESTATUS' as readonly, whether or not it be sensible to attempt
to assign a value to any of them.

Now consider the prior definition of the 'bash_misc_vars' variable,
which was as follows.

BASH BASH_.* COLUMNS COMP_WORDBREAKS HISTCMD HISTFILE HOSTNAME HOSTTYPE
IFS LINENO MACHTYPE OLDPWD OPTERR OPTIND OSTYPE POSIXLY_CORRECT PS4 PWD
RANDOM SECONDS SHLVL _

Putting aside the "BASH_.*" pattern, there are 21 words, which is rather
short of the mark. By my count there are 105 variables in bash-5.3-rc2
that fall into any of the following three categories.

a) initially set by BASH
b) later set by BASH in response to particular events
c) not set by BASH but able to influence its behaviour if set

In fairness, many of these variables are not expected to have been
exported and/or pertain only to the interactive context, though I would
point out that misguided users sometimes export variables that ought not
to be exported. Regardless, I would aver that portage ought to be
thorough and try its level best never to pollute environment dumps with
the variables in question. There are attendant concerns regarding
hygiene, safety, and even privacy in some cases.

Address this issue in a manner twofold. Firstly, by introducing a
routine whereby bash(1) is executed with an empty environment and asked
to report the names of the variables that it has declared. This, alone,
accounts for at least 44 variables. Secondly, revise the set of
hard-coded bash variable names so as to be comprehensive.

With this change, I was immediately able to observe an improvement in
"environment.bz2" hygiene. In particular, the following variables are
now excluded.

- EPOCHREALTIME
- EPOCHSECONDS
- MAIL
- SHELL
- SRANDOM

See-also: 9c80c1b5ee3bfe832a114f33970f0dbad7c3f713
See-also: 5688c04e90e3afd05063a2a4f62930498d8d4dad
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>