isolated-functions.sh: avoid issuing wait -n in parallel() where no jobs are left
Presently, the parallel() function executes a read loop, consuming
null-terminated records from STDIN that are used to compose simple
commands that are executed asychronously. One this loop has concluded, a
second loop begins in which any outstanding subprocesses are waited for.
The latter loop is implemented as an infinite one that breaks upon the
wait -n command producing an exit status of 127, for that is the number
indicating that there are no more subprocesses left to be reaped.

Sam reported that this technique raises an annoying internal debug
message for the bash 5.3 release candidates. As the following examples
show, it can be reproduced by issuing wait -n from the subshell of a
non-interactive shell that has no remaining jobs in flight.

$ bash -c '( : & wait -n; wait -n )'
bash: line 1: DEBUG warning: notify_of_job_status: catch-all setting
J_NOTIFIED on job 0 (32), startup state = 2

$ bash -c '{ : & wait -n; wait -n; } | :'
bash: line 1: DEBUG warning: notify_of_job_status: catch-all setting
J_NOTIFIED on job 0 (32), startup state = 2

It should be noted that such messages only appear in DEBUG builds.
Therefore, there is no chance of them appearing in the upcoming 5.3.0
release. Nevertheless, address this issue by tasking the 'i' variable
with keeping track of how many jobs are in flight, and by having the
second loop iterate exactly as many times as 'i' can be decremented
until it reaches 0.

Reported-by: Sam James <sam@gentoo.org>
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>